requirements.txt
asgiref==1.1.2
asn1crypto==0.24.0
astroid==1.6.2
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
braintree==3.42.0
certifi==2018.1.18
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.1
daphne==1.4.2
Django==2.0.3
django-crispy-forms==1.7.0
django-extensions==2.0.6
django-payments==0.12.1
djangorestframework==3.7.7
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.18
PyJWT==1.6.1
pylint==1.8.3
pytz==2018.3
requests==2.18.4
six==1.11.0
stripe==1.79.1
suds-jurko==0.6
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
wrapt==1.10.11
xmltodict==0.11.0
zope.interface==4.4.3
